Create LCDExample1.0.ino
Code for an LCD display which shows the values of 4 analogue inputs. In this case 2 potentiometers and 2 switches.
The values of each input are displayed two at a time on the LCD (with a 2 second delay).
If a value is changed it will display the changed value "live" on the LCD for 2 seconds after the last change, before reverting back to the main display.
I will now be working on adding a multiplexer to the code with an additional 8 potentiometers.